requirements.txt

viam-sdk==0.42.0

typing-extensions

BAC0==2024.09.20
